requirements.txt
Flask
requests
flask-CORS
mysql-connector-python